All posts mutations locked behind header autentication

+++ src/resolvers/Mutation.js
@@ -22,17 +22,21 @@ const Mutation = {
             token: jwt.sign({ userId: user.id }, 'thisissecret')
         }
     },
-    async deleteUser(parent, args, { prisma }, info) {
+    async deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.deleteUser({
             where: {
-                id: args.id
+                id: userId
             }
         }, info)
     },
-    async updateUser(parent, args, { prisma }, info) {
+    async updateUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.updateUser({
             where: {
-                id: args.id
+                id: userId
             },
             data: args.data
         }, info)
@@ -61,7 +65,7 @@ const Mutation = {
     },
     createPost(parent, args, { prisma, request }, info) {
         const userId = getUserId(request)
-        
+
         return prisma.mutation.createPost({
             data: {
                 title: args.data.title,
@@ -75,7 +79,19 @@ const Mutation = {
             }
         }, info)
     },
-    deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!postExists) {
+            throw new Error('Unable to delete')
+        }
+
         return prisma.mutation.deletePost({
             where: {
                 id: args.id
